improved laser emitter (now has range 32 and goes through blocks based on collision shapes), fixed wrench not rotating laser emitter and machines

+++ src/main/java/net/zaharenko424/a_changed/client/renderer/blockEntity/LaserEmitterRenderer.java
@@ -66,9 +66,9 @@ public class LaserEmitterRenderer implements BlockEntityRenderer<LaserEmitterEnt
         MatrixStack.pop(stack);
     }
 
-    protected void setupBeam(Direction direction, int length){
+    protected void setupBeam(Direction direction, float length){
         beam.zScale = length;
-        float offset = (.5f + (float) length / 2) * 16;
+        float offset = (.5f + length / 2) * 16;
         switch (direction){
             case NORTH -> beam.z -= offset;
             case SOUTH -> beam.z += offset;
